+++ docs-src/src/containers/BodyConfigSectionContainer.jsx
@@ -2,6 +2,7 @@ import { connect } from 'react-redux'
 
 import BodyConfigGeneral from 'components/BodyConfigGeneral'
 import BodyConfigScheduledSleep from 'components/BodyConfigScheduledSleep'
+import BodyConfigScheduledStop from 'components/BodyConfigScheduledStop'
 import BodyConfigExpeditions from 'components/BodyConfigExpeditions'
 import BodyConfigPvP from 'components/BodyConfigPvP'
 import BodyConfigCombat from 'components/BodyConfigCombat'
@@ -17,6 +18,7 @@ const mapStateToProps = state => (
 
 export const BodyConfigGeneralContainer = connect(mapStateToProps)(BodyConfigGeneral)
 export const BodyConfigScheduledSleepContainer = connect(mapStateToProps)(BodyConfigScheduledSleep)
+export const BodyConfigScheduledStopContainer = connect(mapStateToProps)(BodyConfigScheduledStop)
 export const BodyConfigExpeditionsContainer = connect(mapStateToProps)(BodyConfigExpeditions)
 export const BodyConfigPvPContainer = connect(mapStateToProps)(BodyConfigPvP)
 export const BodyConfigCombatContainer = connect(mapStateToProps)(BodyConfigCombat)